ml_service/ml_service/requirements.txt
flask
numpy
pandas
scikit-learn
transformers
opencv-python-headless
pillow
